+++ api/src/main/kotlin/ru/itmo/wisher/api/ApiApplication.kt
@@ -3,11 +3,29 @@ package ru.itmo.wisher.api
 import org.springframework.boot.autoconfigure.SpringBootApplication
 import org.springframework.boot.runApplication
 import org.springframework.boot.web.servlet.ServletComponentScan
+import org.springframework.context.annotation.Bean
+import org.springframework.context.annotation.Configuration
+import org.springframework.web.servlet.config.annotation.CorsRegistry
+import org.springframework.web.servlet.config.annotation.WebMvcConfigurer
 
-@ServletComponentScan
 @SpringBootApplication
 class ApiApplication
 
 fun main(args: Array<String>) {
     runApplication<ApiApplication>(*args)
+}
+
+@Configuration
+class GlobalApplicationConfiguration {
+    @Bean
+    fun corsConfigurer() =
+        object : WebMvcConfigurer {
+            override fun addCorsMappings(registry: CorsRegistry) {
+                registry
+                    .addMapping("/**")
+                    .allowedOriginPatterns("*")
+                    .allowedMethods("*")
+                    .allowedHeaders("*")
+            }
+        }
 }
